sw-refapp-mra-app-infra-code-main-aws/requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.28.52
